refactor: separate answer submission into middleware (#54)
Factor out answer submission logic in router, into middleware.
